% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bfmThresh.R
\name{bfmThresh}
\alias{bfmThresh}
\title{BFM With Magnitude Threshold}
\usage{
bfmThresh(x, dates, start, monend = NULL, magnThresh = 0, magnPeriod = 1,
  adjust = 1, verbose = FALSE, formula = response ~ harmon, order = 1,
  returnBFM = FALSE, ...)
}
\arguments{
\item{x}{Numeric. Time series vector}

\item{dates}{Time (date) index of \code{x}}

\item{start}{Numeric. Start of monitoring period. See \code{\link{bfastmonitor}}}

\item{monend}{Numeric. Optional: end of the monitoring period. See \code{\link{bfmSpatial}}}

\item{magnThresh}{Numeric. Threshold for breakpoints. All breakpoints with magnitude < \code{magnThresh} are ignored, all data \code{adjust} years preceding this breakpoint are deleted, and monitoring continues with a revised history period.}

\item{magnPeriod}{Numeric. Period of time (in years) from which to compute magnitude (as median of residuals)}

\item{adjust}{Numeric. Delete all observations within this many years preceding a breakpoint, if that breakpoint has a magnitude less than \code{magnThresh}}

\item{verbose}{Logical. Display error messages (if any)?}

\item{formula}{See \code{\link{bfastmonitor}}}

\item{order}{See \code{\link{bfastmonitor}}}

\item{returnBFM}{Logical. Return a regular \code{bfastmonitor} object?}

\item{...}{Arguments to pass to \code{\link{bfastmonitor}}}
}
\value{
Numeric vector with output: break date, magnitude, length of history period (in years), R-squared of fitted model, Adjusted R-squared of fitted model, Intercept of fitted model, cosine coefficient(s) (if applicable), sine coefficient(s) (if applicable), error flag (1 means an error was encountered in the \code{try()} statement). If \code{returnBFM} is \code{FALSE}, a list with the above values as first item and a regular BFM object as second item is returned.
}
\description{
bfm with automatic magn thresholding and re-monitoring
}
\details{
The magnitude is computed as the median of the residuals within the period \code{magnPeriod} years after the break date. If this magnitude is greater than \code{magnThresh}, all observations preceding the break date by \code{adjust} years are deleted (including the observation at the break date), the start of the monitoring period is reset to the break date, the model is re-fit and monitoring continues until the end of the time series is reached, or a breakpoint with magnitude > \code{magnThresh} is found.

The default value for \code{magnPeriod} is 1, assuming that no follow-up land use or other dynamics take place within a 1-year period.

The amplitude of the fitted 1st order model can be computed by taking \code{sqrt{harmoncos^2 + harmonsin^2}}
}
\examples{
library(bfmThresh)
data(tura)
s <- getSceneinfo(names(tura))
xy <- c(820692.4, 829564.4)
z <- zoo(tura[cellFromXY(tura, xy)][1, ], s$date)
z <- window(z, start = "1999-01-01")
bfm <- bfmThresh(z, dates = time(z), start = c(2005, 1), plot = TRUE)
print(bfm)
bfm2 <- bfmThresh(z, dates = time(z), start = c(2005, 1), magnThresh = -500, plot = TRUE)
print(bfm2)
}
\author{
Ben DeVries
}

